+++ R/metNOAA.R
@@ -87,7 +87,6 @@
 #'   the chosen location e.g.  `path = "C:/Users/David"`. Files are saved by
 #'   year and site.
 #' @export
-#' @import readr tidyr dplyr
 #' @return Returns a data frame of surface observations. The data frame is
 #'   consistent for use with the `openair` package. Note that the data are
 #'   returned in GMT (UTC) time zone format. Users may wish to express the data
@@ -159,16 +158,16 @@ importNOAA <- function(
   }
 
   # check to see what is missing and print to screen
-  actual <- select(dat, code, date, station) |>
-    mutate(year = as.numeric(format(date, "%Y"))) |>
-    group_by(code, year) |>
-    slice(1)
+  actual <- dplyr::select(dat, code, date, station) |>
+    dplyr::mutate(year = as.numeric(format(date, "%Y"))) |>
+    dplyr::group_by(code, year) |>
+    dplyr::slice(1)
 
-  actual <- left_join(site_process, actual, by = c("code", "year"))
+  actual <- dplyr::left_join(site_process, actual, by = c("code", "year"))
 
   if (length(which(is.na(actual$date))) > 0 && !quiet) {
     print("The following sites / years are missing:")
-    print(filter(actual, is.na(date)))
+    print(dplyr::filter(actual, is.na(date)))
   }
 
   if (!is.na(path)) {
@@ -186,9 +185,9 @@ importNOAA <- function(
       return(dat)
     }
 
-    mutate(dat, year = format(date, "%Y")) |>
-      group_by(code, year) |>
-      do(writeMet(.))
+    dplyr::mutate(dat, year = format(date, "%Y")) |>
+      dplyr::group_by(code, year) |>
+      dplyr::do(writeMet(.))
   }
 
   return(dat)
@@ -215,30 +214,30 @@ getDat <- function(code, year, hourly) {
   # suppress warnings because some fields might be missing in the list
   # Note that not all available data is returned - just what I think is most useful
   met_data <- try(
-    suppressWarnings(read_csv(
+    suppressWarnings(readr::read_csv(
       file.name,
-      col_types = cols_only(
-        STATION = col_character(),
-        DATE = col_datetime(format = ""),
-        SOURCE = col_double(),
-        LATITUDE = col_double(),
-        LONGITUDE = col_double(),
-        ELEVATION = col_double(),
-        NAME = col_character(),
-        REPORT_TYPE = col_character(),
-        CALL_SIGN = col_double(),
-        QUALITY_CONTROL = col_character(),
-        WND = col_character(),
-        CIG = col_character(),
-        VIS = col_character(),
-        TMP = col_character(),
-        DEW = col_character(),
-        SLP = col_character(),
-        AA1 = col_character(),
-        AW1 = col_character(),
-        GA1 = col_character(),
-        GA2 = col_character(),
-        GA3 = col_character()
+      col_types = readr::cols_only(
+        STATION = readr::col_character(),
+        DATE = readr::col_datetime(format = ""),
+        SOURCE = readr::col_double(),
+        LATITUDE = readr::col_double(),
+        LONGITUDE = readr::col_double(),
+        ELEVATION = readr::col_double(),
+        NAME = readr::col_character(),
+        REPORT_TYPE = readr::col_character(),
+        CALL_SIGN = readr::col_double(),
+        QUALITY_CONTROL = readr::col_character(),
+        WND = readr::col_character(),
+        CIG = readr::col_character(),
+        VIS = readr::col_character(),
+        TMP = readr::col_character(),
+        DEW = readr::col_character(),
+        SLP = readr::col_character(),
+        AA1 = readr::col_character(),
+        AW1 = readr::col_character(),
+        GA1 = readr::col_character(),
+        GA2 = readr::col_character(),
+        GA3 = readr::col_character()
       ),
       progress = FALSE
     )),
@@ -251,7 +250,7 @@ getDat <- function(code, year, hourly) {
     return()
   }
 
-  met_data <- rename(
+  met_data <- dplyr::rename(
     met_data,
     code = STATION,
     station = NAME,
@@ -266,9 +265,13 @@ getDat <- function(code, year, hourly) {
   # separate WND column
 
   if ("WND" %in% names(met_data)) {
-    met_data <- separate(met_data, WND, into = c("wd", "x", "y", "ws", "z"))
+    met_data <- tidyr::separate(
+      met_data,
+      WND,
+      into = c("wd", "x", "y", "ws", "z")
+    )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       wd = as.numeric(wd),
       wd = ifelse(wd == 999, NA, wd),
@@ -280,14 +283,14 @@ getDat <- function(code, year, hourly) {
 
   # separate TMP column
   if ("TMP" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       TMP,
       into = c("air_temp", "flag_temp"),
       sep = ","
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       air_temp = as.numeric(air_temp),
       air_temp = ifelse(air_temp == 9999, NA, air_temp),
@@ -297,7 +300,7 @@ getDat <- function(code, year, hourly) {
 
   # separate VIS column
   if ("VIS" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       VIS,
       into = c("visibility", "flag_vis1", "flag_vis2", "flag_vis3"),
@@ -305,7 +308,7 @@ getDat <- function(code, year, hourly) {
       fill = "right"
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       visibility = as.numeric(visibility),
       visibility = ifelse(visibility %in% c(9999, 999999), NA, visibility)
@@ -314,14 +317,14 @@ getDat <- function(code, year, hourly) {
 
   # separate DEW column
   if ("DEW" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       DEW,
       into = c("dew_point", "flag_dew"),
       sep = ","
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       dew_point = as.numeric(dew_point),
       dew_point = ifelse(dew_point == 9999, NA, dew_point),
@@ -330,7 +333,7 @@ getDat <- function(code, year, hourly) {
   }
   # separate SLP column
   if ("SLP" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       SLP,
       into = c("atmos_pres", "flag_pres"),
@@ -338,7 +341,7 @@ getDat <- function(code, year, hourly) {
       fill = "right"
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       atmos_pres = as.numeric(atmos_pres),
       atmos_pres = ifelse(atmos_pres %in% c(99999, 999999), NA, atmos_pres),
@@ -348,7 +351,7 @@ getDat <- function(code, year, hourly) {
 
   # separate CIG (sky condition) column
   if ("CIG" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       CIG,
       into = c("ceil_hgt", "flag_sky1", "flag_sky2", "flag_sky3"),
@@ -356,7 +359,7 @@ getDat <- function(code, year, hourly) {
       fill = "right"
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       ceil_hgt = as.numeric(ceil_hgt),
       ceil_hgt = ifelse(ceil_hgt == 99999, NA, ceil_hgt)
@@ -370,7 +373,7 @@ getDat <- function(code, year, hourly) {
 
   if ("GA1" %in% names(met_data)) {
     # separate GA1 (cloud layer 1 height, amount) column
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       GA1,
       into = c(
@@ -384,7 +387,7 @@ getDat <- function(code, year, hourly) {
       sep = ","
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       cl_1 = as.numeric(cl_1),
       cl_1 = ifelse((is.na(cl_1) & ceil_hgt == 22000), 0, cl_1),
@@ -395,7 +398,7 @@ getDat <- function(code, year, hourly) {
   }
 
   if ("GA2" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       GA2,
       into = c(
@@ -409,7 +412,7 @@ getDat <- function(code, year, hourly) {
       sep = ","
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       cl_2 = as.numeric(cl_2),
       cl_2 = ifelse(cl_2 == 99, NA, cl_2),
@@ -419,7 +422,7 @@ getDat <- function(code, year, hourly) {
   }
 
   if ("GA3" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       GA3,
       into = c(
@@ -433,7 +436,7 @@ getDat <- function(code, year, hourly) {
       sep = ","
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       cl_3 = as.numeric(cl_3),
       cl_3 = ifelse(cl_3 == 99, NA, cl_3),
@@ -454,14 +457,14 @@ getDat <- function(code, year, hourly) {
 
   # PRECIP AA1
   if ("AA1" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       AA1,
       into = c("precip_code", "precip_raw", "code_1", "code_2"),
       sep = ","
     )
 
-    met_data <- mutate(
+    met_data <- dplyr::mutate(
       met_data,
       precip_raw = as.numeric(precip_raw),
       precip_raw = ifelse(precip_raw == 9999, NA, precip_raw),
@@ -487,7 +490,7 @@ getDat <- function(code, year, hourly) {
   # weather codes, AW1
 
   if ("AW1" %in% names(met_data)) {
-    met_data <- separate(
+    met_data <- tidyr::separate(
       met_data,
       AW1,
       into = c("pwc", "code_1"),
@@ -495,15 +498,15 @@ getDat <- function(code, year, hourly) {
       fill = "right"
     )
 
-    met_data <- left_join(met_data, worldmet::weatherCodes, by = "pwc")
-    met_data <- select(met_data, -pwc) |>
-      rename(pwc = description)
+    met_data <- dplyr::left_join(met_data, worldmet::weatherCodes, by = "pwc")
+    met_data <- dplyr::select(met_data, -pwc) |>
+      dplyr::rename(pwc = description)
   }
 
   ## select the variables we want
-  met_data <- select(
+  met_data <- dplyr::select(
     met_data,
-    any_of(c(
+    dplyr::any_of(c(
       "date",
       "code",
       "station",
@@ -555,7 +558,7 @@ getDat <- function(code, year, hourly) {
 
   ## add pwc back in
   if (exists("pwc")) {
-    met_data <- left_join(met_data, pwc, by = "date")
+    met_data <- dplyr::left_join(met_data, pwc, by = "date")
   }
 
   ## add precipitation - based on 12 HOUR averages, so work with hourly data
@@ -586,5 +589,5 @@ getDat <- function(code, year, hourly) {
     replace(x, is.nan(x), NA)
   })
 
-  return(as_tibble(met_data))
+  return(dplyr::as_tibble(met_data))
 }