+++ R/worldmet-package.R
@@ -20,6 +20,6 @@
 "_PACKAGE"
 ## usethis namespace: start
 #' @importFrom foreach %dopar%
-#' @importFrom tibble tibble
+#' @importFrom rlang .data
 ## usethis namespace: end
 NULL